backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
python-dotenv==1.0.0
pydantic==2.5.2
google-generativeai==0.3.1
python-multipart==0.0.6 